// mesh-control-plane/mesh-ui/src/api/openapi/types.gen.ts
export type V1DeleteNodeResponse = {
  [key: string]: unknown
}

export type V1DeleteUserResponse = {
  [key: string]: unknown
}

export type V1ExpireApiKeyRequest = {
  prefix?: string
}

export type V1ExpireApiKeyResponse = {
  [key: string]: unknown
}

export type V1ExpireNodeResponse = {
  node?: V1Node
}

export type V1ExpirePreAuthKeyRequest = {
  user?: string
  key?: string
}

export type V1ExpirePreAuthKeyResponse = {
  [key: string]: unknown
}

export type V1GetNodeResponse = {
  node?: V1Node
}

export type V1GetPolicyResponse = {
  policy?: string
  updatedAt?: string
}

export type V1ListApiKeysResponse = {
  apiKeys?: Array<V1ApiKey>
}

export type V1ListNodesResponse = {
  nodes?: Array<V1Node>
}

export type V1ListPreAuthKeysResponse = {
  preAuthKeys?: Array<V1PreAuthKey>
}

export type V1ListUsersResponse = {
  users?: Array<V1User>
}

export type V1MoveNodeResponse = {
  node?: V1Node
}

export type V1Node = {
  id?: string
  machineKey?: string
  nodeKey?: string
  discoKey?: string
  ipAddresses?: Array<string>
  name?: string
  user?: V1User
  lastSeen?: string
  expiry?: string
  preAuthKey?: V1PreAuthKey
  createdAt?: string
  registerMethod?: V1RegisterMethod
  forcedTags?: Array<string>
  invalidTags?: Array<string>
  validTags?: Array<string>
  givenName?: string
  online?: boolean
  approvedRoutes?: Array<string>
  availableRoutes?: Array<string>
  subnetRoutes?: Array<string>
}

export type V1PreAuthKey = {
  user?: V1User
  id?: string
  key?: string
  reusable?: boolean
  ephemeral?: boolean
  used?: boolean
  expiration?: string
  createdAt?: string
  aclTags?: Array<string>
}

export type V1RegisterMethod =
  | 'REGISTER_METHOD_UNSPECIFIED'
  | 'REGISTER_METHOD_AUTH_KEY'
  | 'REGISTER_METHOD_CLI'
  | 'REGISTER_METHOD_OIDC'

export type V1RegisterNodeResponse = {
  node?: V1Node
}

export type V1RenameNodeResponse = {
  node?: V1Node
}

export type V1RenameUserResponse = {
  user?: V1User
}

export type V1SetApprovedRoutesResponse = {
  node?: V1Node
}

export type V1SetPolicyRequest = {
  policy?: string
}

export type V1SetPolicyResponse = {
  policy?: string
  updatedAt?: string
}

export type V1SetTagsResponse = {
  node?: V1Node
}

export type V1User = {
  id?: string
  name?: string
  createdAt?: string
  displayName?: string
  email?: string
  providerId?: string
  provider?: string
  profilePicUrl?: string
}